Convert to gradle project
