+++ build.gradle
@@ -15,4 +15,5 @@ dependencies {
     compileOnly 'org.spigotmc:spigot-api:1.14.4-R0.1-SNAPSHOT'
     compileOnly 'com.github.SkriptLang:Skript:2.3'
     compileOnly 'me.clip:placeholderapi:2.10.4'
+    compileOnly fileTree(dir: 'libs', include: '*.jar')
 }
